refactor: separate home shell from MIDI project list
