+++ src/components/midi-project-list.tsx
@@ -1,11 +1,5 @@
 import { useMutation } from "@tanstack/react-query";
-import {
-  AudioLinesIcon,
-  GitForkIcon,
-  Music2Icon,
-  Pencil,
-  Trash2,
-} from "lucide-react";
+import { Pencil, Trash2 } from "lucide-react";
 import { useEffect, useRef, useState } from "react";
 import { toast } from "sonner";
 import { useDraftTextInput } from "../hooks/use-draft-text-input";
@@ -14,135 +8,9 @@ import { parseProjectFile } from "../lib/project-file";
 import { type ProjectMetadata, projectStorage } from "../lib/project-storage";
 import { routes } from "../lib/routes";
 import { FileDropInput } from "./file-drop-input";
-import { RecorderProjectList } from "./recorder/project-list";
 import { Button } from "./ui/button";
 
-type ProjectType = "midi" | "recorder";
-
-export function ProjectListView() {
-  const [projectType, setProjectType] = useState<ProjectType>(
-    () => projectStorage.readPreferences().projectType,
-  );
-
-  const selectProjectType = (type: ProjectType) => {
-    projectStorage.updatePreferences({ projectType: type });
-    setProjectType(type);
-  };
-
-  return (
-    <div
-      data-testid="startup-screen"
-      className="fixed inset-0 z-50 overflow-hidden bg-neutral-900"
-    >
-      {/* Gradient glow */}
-      <div className="pointer-events-none absolute inset-x-0 top-0 h-[28rem] bg-[radial-gradient(ellipse_70%_70%_at_50%_0%,#10b9811f_0%,transparent_70%)]" />
-
-      <div className="relative mx-auto flex h-full w-full max-w-4xl flex-col px-8 py-12">
-        <header className="flex items-start justify-between">
-          <div>
-            <h1 className="text-3xl font-bold text-neutral-100 tracking-tight">
-              Toy MIDI
-            </h1>
-            <p className="mt-1 text-sm text-neutral-500">
-              Edit MIDI and record audio
-            </p>
-          </div>
-          <nav className="flex items-center gap-4 text-sm text-neutral-500">
-            <a
-              href={routes.latencyChecker.href()}
-              className="inline-flex items-center gap-1.5 hover:text-emerald-400 transition-colors"
-            >
-              <AudioLinesIcon className="size-4" />
-              Latency Checker
-            </a>
-            <a
-              href={routes.scoreViewer.href()}
-              data-testid="score-viewer-link"
-              className="inline-flex items-center gap-1.5 hover:text-emerald-400 transition-colors"
-            >
-              <Music2Icon className="size-4" />
-              Score Viewer
-            </a>
-            <a
-              href="https://github.com/hi-ogawa/toy-midi/"
-              target="_blank"
-              rel="noreferrer"
-              className="inline-flex items-center gap-1.5 hover:text-emerald-400 transition-colors"
-            >
-              <GitForkIcon className="size-4" />
-              GitHub
-            </a>
-          </nav>
-        </header>
-
-        <main className="mt-14 min-h-0 flex-1">
-          <h2 className="mb-4 text-xs font-semibold uppercase tracking-[0.16em] text-neutral-400">
-            Your Projects
-          </h2>
-          <div
-            role="tablist"
-            aria-label="Project type"
-            className="mb-4 flex gap-2"
-          >
-            {(["midi", "recorder"] as const).map((type) => (
-              <button
-                key={type}
-                type="button"
-                role="tab"
-                id={`project-tab-${type}`}
-                aria-controls="project-panel"
-                aria-selected={projectType === type}
-                tabIndex={projectType === type ? 0 : -1}
-                onClick={() => selectProjectType(type)}
-                onKeyDown={(event) => {
-                  let nextType: ProjectType;
-                  switch (event.key) {
-                    case "ArrowLeft":
-                    case "ArrowRight": {
-                      nextType = type === "midi" ? "recorder" : "midi";
-                      break;
-                    }
-                    case "Home": {
-                      nextType = "midi";
-                      break;
-                    }
-                    case "End": {
-                      nextType = "recorder";
-                      break;
-                    }
-                    default: {
-                      return;
-                    }
-                  }
-                  event.preventDefault();
-                  selectProjectType(nextType);
-                  document.getElementById(`project-tab-${nextType}`)?.focus();
-                }}
-                className="rounded-md px-4 py-2 text-sm text-neutral-400 hover:bg-neutral-800 aria-selected:bg-neutral-700 aria-selected:text-neutral-100"
-              >
-                {type === "midi" ? "MIDI" : "Recorder"}
-              </button>
-            ))}
-          </div>
-          <div
-            role="tabpanel"
-            id="project-panel"
-            aria-labelledby={`project-tab-${projectType}`}
-            tabIndex={0}
-          >
-            {projectType === "midi" ? (
-              <MidiProjectList />
-            ) : (
-              <RecorderProjectList />
-            )}
-          </div>
-        </main>
-      </div>
-    </div>
-  );
-}
-
-function MidiProjectList() {
+export function MidiProjectList() {
   const [renamingProjectId, setRenamingProjectId] = useState<string>();
   const [projects, setProjects] = useState(projectStorage.listMetadata());
